+++ Pandora/Pattern/Functor/Applicative.hs
@@ -1,7 +1,8 @@
 module Pandora.Pattern.Functor.Applicative (Applicative (..)) where
 
+import Pandora.Core.Functor (type (:.:))
 import Pandora.Core.Morphism (identity)
-import Pandora.Pattern.Functor.Covariant (Covariant ((<$)))
+import Pandora.Pattern.Functor.Covariant (Covariant ((<$>), (<$)))
 
 infixl 4 <*>, <*, *>
 
@@ -29,3 +30,15 @@ class Covariant t => Applicative t where
 	-- | Repeat an action indefinitely
 	forever :: t a -> t b
 	forever x = x *> forever x
+
+	-- | Infix versions of `apply` with various nesting levels
+	(<**>) :: Applicative u => (t :.: u) (a -> b) -> (t :.: u) a -> (t :.: u) b
+	f <**> x = (<*>) <$> f <*> x
+	(<***>) :: (Applicative u, Applicative v) => (t :.: u :.: v) (a -> b)
+		-> (t :.: u :.: v) a -> (t :.: u :.: v) b
+	f <***> x = (<**>) <$> f <*> x
+	(<****>) :: (Applicative u, Applicative v, Applicative w)
+		=> (t :.: u :.: v :.: w) (a -> b)
+		-> (t :.: u :.: v :.: w) a
+		-> (t :.: u :.: v :.: w) b
+	f <****> x = (<***>) <$> f <*> x