+++ Pandora/Paradigm/Inventory/Storage.hs
@@ -1,6 +1,6 @@
-module Pandora.Paradigm.Inventory.Storage (Storage (..), Store, position, access, retrofit) where
+module Pandora.Paradigm.Inventory.Storage (Storage (..), position, access, retrofit) where
 
-import Pandora.Core.Functor (type (:.:))
+import Pandora.Core.Functor (type (:.:), type (><))
 import Pandora.Core.Morphism ((.), ($), (?))
 import Pandora.Paradigm.Basis.Identity (Identity)
 import Pandora.Paradigm.Basis.Product (Product ((:*:)), type (:*:))
@@ -10,31 +10,25 @@ import Pandora.Pattern.Functor.Extendable (Extendable ((=>>)))
 import Pandora.Pattern.Functor.Applicative (Applicative ((<*>)))
 import Pandora.Pattern.Functor.Comonad (Comonad)
 
-newtype Storage p t a = Storage { stored :: ((:*:) p :.: t :.: (->) p) a }
+newtype Storage p a = Storage { stored :: (:*:) p :.: (->) p >< a }
 
-instance Covariant t => Covariant (Storage p t) where
-	f <$> Storage (p :*: x) = Storage . (:*:) p $ (f .) <$> x
+instance Covariant (Storage p) where
+	g <$> Storage (p :*: f) = Storage . (:*:) p $ (g .) f
 
-instance Extractable t => Extractable (Storage p t) where
-	extract (Storage (p :*: x)) = extract x p
+instance Extractable (Storage p) where
+	extract (Storage (p :*: x)) = x p
 
-instance Extendable t => Extendable (Storage p t) where
-	Storage (old :*: x) =>> f = Storage . (:*:) old . (=>>) x $
-		\y -> \new -> f . Storage $ new :*: y
+instance Extendable (Storage p) where
+	Storage (old :*: f) =>> g = Storage . (:*:) old
+		$ \new -> g . Storage $ new :*: f
 
-instance Applicative t => Applicative (Storage p t) where
-	Storage (_ :*: x) <*> Storage (q :*: y) = Storage . (:*:) q $
-		(\f g x' -> f x' (g x')) <$> x <*> y
+instance Comonad (Storage p) where
 
-instance Comonad g => Comonad (Storage p g) where
-
-type Store p = Storage p Identity
-
-position :: Storage p t a -> p
+position :: Storage p a -> p
 position (Storage (p :*: _)) = p
 
-access :: Extractable t => p -> Storage p t a -> a
-access p = extract ? p . extract . stored
+access :: p -> Storage p a -> a
+access p = extract ? p . stored
 
-retrofit :: Extractable t => (p -> p) -> Storage p t a -> Storage p t a
-retrofit f (Storage (p :*: x)) = Storage $ (f p) :*: x
+retrofit :: (p -> p) -> Storage p a -> Storage p a
+retrofit g (Storage (p :*: f)) = Storage $ g p :*: f