+++ Pandora/Paradigm/Junction/Transformer.hs
@@ -1,6 +1,6 @@
-module Pandora.Paradigm.Junction.Transformer (T (..), type (:!:), up, Y (..), type (:>:)) where
+module Pandora.Paradigm.Junction.Transformer (UT (..), type (:!:), up, Y (..), type (:>:)) where
 
-import Pandora.Core.Functor (type (:.:))
+import Pandora.Core.Functor (type (:.:), type (><))
 import Pandora.Core.Morphism ((.), ($))
 import Pandora.Pattern.Functor.Covariant (Covariant ((<$>), (<$$>), comap))
 import Pandora.Pattern.Functor.Pointable (Pointable (point))
@@ -18,59 +18,58 @@ import Pandora.Pattern.Object.Semigroup (Semigroup ((+)))
 import Pandora.Pattern.Object.Monoid (Monoid (zero))
 
 infixr 0 :!:, :>:
-type (:!:) t u = T t u
+type (:!:) t u = UT t u
 type (:>:) t u = Y t u
 
+newtype UT t u a = UT { ut :: u :.: t >< a }
 
-newtype T t u a = T { t :: (u :.: t) a }
+instance (Covariant t, Covariant u) => Covariant (UT t u) where
+	f <$> UT x = UT $ f <$$> x
 
-instance (Covariant t, Covariant u) => Covariant (T t u) where
-	f <$> T x = T $ f <$$> x
+instance (Pointable t, Pointable u) => Pointable (UT t u) where
+	point = UT . point . point
 
-instance (Pointable t, Pointable u) => Pointable (T t u) where
-	point = T . point . point
+instance (Extractable t, Extractable u) => Extractable (UT t u) where
+	extract = extract . extract . ut
 
-instance (Extractable t, Extractable u) => Extractable (T t u) where
-	extract = extract . extract . t
+instance (Covariant t, Avoidable u) => Avoidable (UT t u) where
+	idle = UT idle
 
-instance (Covariant t, Avoidable u) => Avoidable (T t u) where
-	idle = T idle
+instance (Covariant t, Alternative u) => Alternative (UT t u) where
+	UT x <+> UT y = UT $ x <+> y
 
-instance (Covariant t, Alternative u) => Alternative (T t u) where
-	T x <+> T y = T $ x <+> y
+instance (Applicative t, Applicative u) => Applicative (UT t u) where
+	UT f <*> UT x = UT $ apply <$> f <*> x
 
-instance (Applicative t, Applicative u) => Applicative (T t u) where
-	T f <*> T x = T $ apply <$> f <*> x
+instance Pointable t => Liftable (UT t) where
+	lift x = UT $ point <$> x
 
-instance Pointable t => Liftable (T t) where
-	lift x = T $ point <$> x
+instance Extractable t => Lowerable (UT t) where
+	lower (UT x) = extract <$> x
 
-instance Extractable t => Lowerable (T t) where
-	lower (T x) = extract <$> x
+instance (Traversable t, Traversable u) => Traversable (UT t u) where
+	UT x ->> f = UT <$> x ->>> f
 
-instance (Traversable t, Traversable u) => Traversable (T t u) where
-	T x ->> f = T <$> x ->>> f
+instance (Distributive t, Distributive u) => Distributive (UT t u) where
+	x >>- f = UT . comap distribute . distribute $ ut . f <$> x
 
-instance (Distributive t, Distributive u) => Distributive (T t u) where
-	x >>- f = T . comap distribute . distribute $ t . f <$> x
+instance Setoid (u :.: t >< a) => Setoid (UT t u a) where
+	UT x == UT y = x == y
 
-instance Setoid ((u :.: t) a) => Setoid (T t u a) where
-	T x == T y = x == y
+instance Chain (u :.: t >< a) => Chain (UT t u a) where
+	UT x <=> UT y = x <=> y
 
-instance Chain ((u :.: t) a) => Chain (T t u a) where
-	T x <=> T y = x <=> y
+instance Semigroup (u :.: t >< a) => Semigroup (UT t u a) where
+	UT x + UT y = UT $ x + y
 
-instance Semigroup ((u :.: t) a) => Semigroup (T t u a) where
-	T x + T y = T $ x + y
+instance Monoid (u :.: t >< a) => Monoid (UT t u a) where
+	zero = UT zero
 
-instance Monoid ((u :.: t) a) => Monoid (T t u a) where
-	zero = T zero
-
-up :: Pointable u => t a -> T t u a
-up = T . point
+up :: Pointable u => t a -> UT t u a
+up = UT . point
 
 
-newtype Y t u a = Y { y :: (u :.: t u) a }
+newtype Y t u a = Y { y :: u :.: t u >< a }
 
 instance (Covariant (t u), Covariant u) => Covariant (Y t u) where
 	f <$> Y x = Y $ f <$$> x
@@ -102,14 +101,14 @@ instance (forall u' . Pointable u', Liftable t) => Liftable (Y t) where
 instance (forall u' . Extractable u', Lowerable t) => Lowerable (Y t) where
 	lower = lower . extract . y
 
-instance (forall u' . Setoid ((u' :.: t u') a)) => Setoid (Y t u a) where
+instance (forall u' . Setoid (u' :.: t u' >< a)) => Setoid (Y t u a) where
 	Y x == Y y = x == y
 
-instance (forall u' . Chain ((u' :.: t u') a)) => Chain (Y t u a) where
+instance (forall u' . Chain (u' :.: t u' >< a)) => Chain (Y t u a) where
 	Y x <=> Y y = x <=> y
 
-instance (forall u' . Semigroup ((u' :.: t u') a)) => Semigroup (Y t u a) where
+instance (forall u' . Semigroup (u' :.: t u' >< a)) => Semigroup (Y t u a) where
 	Y x + Y y = Y $ x + y
 
-instance (forall u' . Monoid ((u' :.: t u') a)) => Monoid (Y t u a) where
+instance (forall u' . Monoid (u' :.: t u' >< a)) => Monoid (Y t u a) where
 	zero = Y zero