+++ CHANGELOG.md
@@ -65,3 +65,4 @@
 * Change `Stack` and `Nonempty` definitions, temporarily remove `filter`
 * Define `<**>`, `<***>`, `<****>` methods for `Applicative` to compact expressions
 * Change `Graph` definition
+* Rename `ask` to `env` method of `Environmental` datatype