+++ CHANGELOG.md
@@ -66,3 +66,4 @@
 * Define `<**>`, `<***>`, `<****>` methods for `Applicative` to compact expressions
 * Change `Graph` definition
 * Rename `ask` to `env` method of `Environmental` datatype
+* Introduce `><` type operator to separate functors from its arguments