+++ CHANGELOG.md
@@ -69,3 +69,4 @@
 * Introduce `><` type operator to separate functors from its arguments
 * Define `Determinable` typeclass and define its instance for `Predicate`
 * Define `curry` and `uncurry` for `Product` datatype
+* Flip arguments of `statefully` method of `Stateful` datatype